Added example how to play mp3 file from a standalone java application.
The code should also run in an OSGi environment but there is still
something to fix before that will work appropriately. This is the
purpose for setting up this example project.

My understanding is that OSGi and java SPIs do not play nicely
together. I need to figure out how to get the mp3spi extention
to work within the OSGi container. Apache Aries describes both
the problem and the solution at:

http://aries.apache.org/modules/spi-fly.html

I haven't been able to get the integration to work yet. I'll play
around with it some more and if I still need help then I'll ask
on stackoverflow and the Apache Aries user list.